fix: clean up modules
