+++ crates/bin/pindexer/src/lib.rs
@@ -2,7 +2,6 @@ pub use cometindex::{AppView, Indexer};
 
 mod indexer_ext;
 pub use indexer_ext::IndexerExt;
-
 pub mod block;
 pub mod shielded_pool;
 pub mod stake;
